Fixes #45
--ESDRecordMaker.cs
Added class that holds data for a single maker/manufacturer. A maker/manufacturer builds and/or assembles models which contain a collection of products or raw materials.

--ESDRecordMakerModel.cs
Added Record class that holds data for a single model created or assembled by a maker/manufacturer. Each model is made up of a collection of products or raw materials.

--ESDRecordMakerModelMapping.cs
Added Record class that holds data for a single mapping between a maker/manufacturer's model, a category, and either a product, download, or labour entity. Each mapping may also contain any number of attribute values to contain more data about the mapped entity in relation to the model.

--ESDocumentMaker.cs
Ecommerce standards document class that contains a list of maker/manufacturer records

--ESDocumentMakerModel.cs
Document class that contains a list of model records associated to makers/manufacturers

--ESDocumentMakerModelMapping.cs
Document class that contains a list of maker model mappings records that associate product, download, or labour records to categories within models manufactured or assembled by makers/manufacturers