Fixes #47
--ESDRecordCustomerAccountEnquiryQuote.cs
Added record class that contains the data of a single quote that can be queried for a customer account. A quote record contains information about items that need to be confirmed and priced for a customer account. The account enquiry aspect denotes that the record may be queried in real time and contain additional information associated with the record.

--ESDRecordCustomerAccountEnquiryQuoteLine.cs
Added record class holding data for a customer account quote line. The line stores information about a single type of product, goods or service, and the pricing quoted to purchase the item

--ESDocumentCustomerAccountEnquiry.cs
Added ability to store an array of quoteRecords against a Customer Account Enquiry document, as well as showing a coding example of the serialised form of quote records appearing within the document.