# Delivery Notice document and record classes
--ESDocumentDeliveryNotice.cs
Added class to store a list of delivery notice records.

--ESDRecordDeliveryNotice.cs
Added class to store the details of a single delivery notice, that contains the status of products being delivered.

--ESDRecordDeliveryNoticeLine.cs
Added class to store the details of a single line associated to a delivery notice. The line contains the details of the product being delivered, and which invoice and purchase order it is associated against.

--ESDocumentConstants.cs
Added constants for setting the state of a delivery notice, as well as invoice payments and line types.