Fixes #71
--ESDRecordImage.cs
Added keyCategoryTreeID property to allow category trees to have image records assigned to them.
Added isPrimary property to set if an image is the primary image for a record when multiple images are assigned to the one entity.
Added ordering property to allow image records to be ordered when multiple images are assigned to one entity.