Fixes #75
--ESDRecordSellUnit.cs
Added measurement properties to store information about dimensions and weight of a sell unit and its packaging dimensions.
Added properties to indicate the minimum quantity that a sell unit can be sold as, and the incrementing quantity that determines the quantity of a sell unit can be divided by to allow it to be bought in steps as.

--ESDocumentProduct.cs