Fixes #76
--ESDRecordDeliveryNotice.cs
Added keyCustomerAccountID and keySupplierAccountID properties to link customer accounts and supplier account records to delivery notices.
Added cancelledDate property to store the date when the delivery was cancelled.

--ESDRecordDeliveryNoticeLine.cs
Added keySalesOrderID and keyPurchaseOrderID proeprties to link sales orders and purchase orders to the delivery lines of the records the goods may have been originally ordered from.

--ESDocumentConstants.cs
Added CANCELLED and AWAITING_RECEIVER_PICKUP delivery status to handle if a delivery was cancelled or if it needs to be picked up by the receiver.

--ESDocumentDeliveryNotice.cs
Added additional properties to example JSON delivery notice data.