Fixes #84
--ESDRecordContact.cs
Added keyLocationID property to link a location to a contact.
Added isPrimary property to indicate if a contact is a primary contact in a list of contacts.
Added allowMarketing property to indicate if the contact is allowed to receive marketing material and communications.
Added marketingCategoryCodes property to indicate the categories of marketing material and communications that the contact is interested in receiving.

--ESDocumentLocationContact.cs
Added document class to store a list of contacts records that associated to locations.

--ESDocumentCustomerAccountContact.cs
--ESDocumentSupplierAccountContact.cs
Updated JSON contact examples for customer account and supplier account contacts to show usages of the additional contact properties defined above.